backend/requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==4.2.11
django-compat==1.0.15
django-cors-headers==4.3.1
django4-background-tasks==1.2.9
djangorestframework==3.15.1
idna==3.7
paramiko==3.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyNaCl==1.5.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
